Clarification
In this upload, I clarify why all measurements are dependent on the speed of light. Thus, the attempt to measure the speed of light with a clock is effectively attempting to measure the speed of light against itself so that even if the speed of light dilates, you will always measure c locally.